Decouple the keyspace group key path from TSO
