+++ server/apiv2/router.go
@@ -63,6 +63,6 @@ func NewV2Handler(_ context.Context, svr *server.Server) (http.Handler, apiutil.
 	router.Use(middlewares.Redirector())
 	root := router.Group(apiV2Prefix)
 	handlers.RegisterKeyspace(root)
-	handlers.RegisterTSOKeyspaceGroup(root)
+	handlers.RegisterKeyspaceGroup(root)
 	return router, group, nil
 }
